Fix unreliable webhook triggers

+++ scripts/e2e-test.sh
@@ -20,16 +20,23 @@ export OPERATOR_SDK_VERSION="v0.12.0"
 
 usage() {
     cat <<EOT
-e2e-test.sh [-e|--env=(minikube|minishift)] [-p|--pause]
+e2e-test.sh [-e|--env=(minikube|minishift)] [-p|--pause] [-n|--noimages] 
+            [-d|--nodeployment] [-r|--run=<test name>]
 
 Execute the end-to-end tests on a local minikube or minishift.
 
 -e|--env=(minikube|minishift) sets the environment the tests will be run under
 -p|--pause Pauses after each test step to help with debugging
+-d|--nodeployment Skip the deployment of the operator
+-n|--noimages Skip building the images
+-r|--run=<test name> Only run the specified Go e2e test. This will skip all others.
 
 You can also specify the settings via environment variables (command line parameters take precedence).
 EUNOMIA_TEST_ENV=(minikube|minishift)
 EUNOMIA_TEST_PAUSE=yes
+EUNOMIA_TEST_SKIP_IMAGES=yes
+EUNOMIA_TEST_SKIP_DEPLOYMENT=yes
+EUNOMIA_TEST_GO_RUN=TestGitHubWebhook
 
 EOT
 }
@@ -124,7 +131,24 @@ while (("$#")); do
         ;;
     -e | --env) # set the test environment
         if [ -n "$2" ] && [ "${2:0:1}" != "-" ]; then
-            EUNOMIA_TEST_ENV=$2
+            export EUNOMIA_TEST_ENV=$2
+            shift 2
+        else
+            echo "Error: Argument for $1 is missing" >&2
+            exit 1
+        fi
+        ;;
+    -n | --noimages) # skip building images
+        export EUNOMIA_TEST_SKIP_IMAGES=yes
+        shift
+        ;;
+    -d | --nodeployment) # skip deploying the operator
+        export EUNOMIA_TEST_SKIP_DEPLOYMENT=yes
+        shift
+        ;;
+    -r | --run) # run specific go e2e test
+        if [ -n "$2" ] && [ "${2:0:1}" != "-" ]; then
+            export EUNOMIA_TEST_GO_RUN=$2
             shift 2
         else
             echo "Error: Argument for $1 is missing" >&2
@@ -152,6 +176,8 @@ eval set -- "$PARAMS"
 # Default settings
 export EUNOMIA_TEST_ENV=${EUNOMIA_TEST_ENV:-minikube}
 export EUNOMIA_TEST_PAUSE=${EUNOMIA_TEST_PAUSE:-no}
+export EUNOMIA_TEST_SKIP_IMAGES=${EUNOMIA_TEST_SKIP_IMAGES:-no}
+export EUNOMIA_TEST_SKIP_DEPLOYMENT=${EUNOMIA_TEST_SKIP_DEPLOYMENT:-no}
 
 case "${EUNOMIA_TEST_ENV:-}" in
 minikube) ;;
@@ -207,8 +233,10 @@ elif [[ "${EUNOMIA_TEST_ENV}" == "minishift" ]]; then
 fi
 
 # Ensure clean workspace
-if [[ $(kubectl get namespace $OPERATOR_NAMESPACE) ]]; then
-    kubectl delete namespace $OPERATOR_NAMESPACE
+if [[ "${EUNOMIA_TEST_SKIP_DEPLOYMENT:-}" == "no" ]]; then
+    if [[ $(kubectl get namespace $OPERATOR_NAMESPACE) ]]; then
+        kubectl delete namespace $OPERATOR_NAMESPACE
+    fi
 fi
 
 # Pre-populate the Docker registry in minikube/minishift with images built from the current commit
@@ -218,7 +246,10 @@ if [[ "${EUNOMIA_TEST_ENV}" == "minikube" ]]; then
 elif [[ "${EUNOMIA_TEST_ENV}" == "minishift" ]]; then
     eval "$(minishift docker-env)"
 fi
-GOOS=linux make e2e-test-images
+
+if [[ "${EUNOMIA_TEST_SKIP_IMAGES:-}" == "no" ]]; then
+    GOOS=linux make e2e-test-images
+fi
 
 # Get minikube/minishift IP address
 # shellcheck disable=SC2155
@@ -233,32 +264,42 @@ fi
 # livenessProbe and readinessProbe ports in deploy/helm/eunomia-operator/templates/deployment.yaml
 export OPERATOR_WEBHOOK_PORT=8080
 
-echo "Installing Eunomia Operator"
-# Eunomia setup
-helm template deploy/helm/eunomia-operator/ \
-    --set eunomia.operator.image.tag=dev \
-    --set eunomia.operator.image.pullPolicy=Never \
-    --set eunomia.operator.namespace=$OPERATOR_NAMESPACE | kubectl apply -f -
+if [[ "${EUNOMIA_TEST_SKIP_DEPLOYMENT:-}" == "no" ]]; then
+    echo "Installing Eunomia Operator"
+    # Eunomia setup
+    helm template deploy/helm/eunomia-operator/ \
+        --set eunomia.operator.image.tag=dev \
+        --set eunomia.operator.image.pullPolicy=Never \
+        --set eunomia.operator.namespace=$OPERATOR_NAMESPACE | kubectl apply -f -
 
-# Deployment test
-if kubectl rollout status deployment/eunomia-operator -n $OPERATOR_NAMESPACE; then
-    echo "Eunomia deployment successful"
-else
-    echo "Eunomia deployment failed"
-    exit 1
+    # Deployment test
+    if kubectl rollout status deployment/eunomia-operator -n $OPERATOR_NAMESPACE; then
+        echo "Eunomia deployment successful"
+    else
+        echo "Eunomia deployment failed"
+        exit 1
+    fi
+
+    pause
 fi
 
-pause
+if [ -n "${EUNOMIA_TEST_GO_RUN:-}" ]; then
+    export EXTRA_ARGS="-run ${EUNOMIA_TEST_GO_RUN}"
+    export TIMEOUT="5m"
+else
+    export EXTRA_ARGS=""
+    export TIMEOUT="40m"
+fi
 
 # End-to-end tests
 operator-sdk test local ./test/e2e \
     --namespaced-manifest /dev/null \
     --global-manifest /dev/null \
     --verbose \
-    --go-test-flags "-tags e2e -timeout 40m"
+    --go-test-flags "-tags e2e -timeout ${TIMEOUT} ${EXTRA_ARGS}"
 pause
 
-## Testing hello-world-yaml example
+# Testing hello-world-yaml example
 # Create new namespace
 kubectl create namespace eunomia-hello-world-yaml-demo
 # Create new service account for the runners
